% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{cor_matrix}
\usage{
cor_matrix(year)
}
\arguments{
\item{year}{Specific year to find a correlation matrix}
}
\description{
This function finds a correlation matrix for numeric values for a given year
}
\examples{
cor_matrix(1997)
}
