% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation Function}
\usage{
correlation(year)
}
\arguments{
\item{year}{The year of the data}
}
\description{
This function outputs a correlation matrix for all numeric variables in the data for that year.
}
\examples{
correlation(year)
}
\keyword{correlation}
