% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_numeric_var.R
\name{correlation_numeric_var}
\alias{correlation_numeric_var}
\title{A Correlation Function}
\usage{
correlation_numeric_var(year)
}
\arguments{
\item{Year}{Which year data do you want to check? Defaults to Null}
}
\description{
This function allows you to check the correlation of all numeric continuous variables in one specific year
}
