% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1}
\alias{F1}
\title{Function one (F1 Function)}
\usage{
F1(year)
}
\description{
Your input for this function should be an individual year. 
Your output should be the player for that year who scored
the most points.
}
