% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2.R
\name{F2}
\alias{F2}
\title{Function two (F2 Function)}
\usage{
F2(year)
}
\description{
Your input for this function should be an individual year. 
Your output should be the oldest player in the NBA data
for that year.
}
