% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3.R
\name{F3}
\alias{F3}
\title{Function three (F3 Function)}
\usage{
F3(year)
}
\description{
Your input for this function should be an individual year. 
Your function should limit the data for that year to include
numeric continuous variables only (not categorical data). 
Your output should be the correlation matrix for all numeric
variables in the data for that year.
}
