% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_over_50.R
\name{games_over_50}
\alias{games_over_50}
\title{A Games Function}
\usage{
games_over_50(year)
}
\arguments{
\item{Year}{Which year data do you want to check? Defaults to Null}
}
\description{
This function allows you to check the data of players having attended over 50 games in one specific year
}
