\name{homework5-package}
\alias{homework5-package}
\alias{homework5}
\docType{package}
\title{
\packageTitle{homework5}
}
\description{
\packageDescription{homework5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{homework5}
\packageIndices{homework5}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{homework5}

Maintainer: \packageMaintainer{homework5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
