\name{hw5-package}
\alias{hw5-package}
\alias{hw5}
\docType{package}
\title{
\packageTitle{hw5}
}
\description{
\packageDescription{hw5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hw5}
\packageIndices{hw5}
The only function is randletter. It returns a random letter.
}
\author{
\packageAuthor{hw5}

Maintainer: \packageMaintainer{hw5}
}

\keyword{ package }
