% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{length,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Get Length of sparse_numeric}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
Integer length.
}
\description{
Get Length of sparse_numeric
}
