% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of sparse_numeric}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
Numeric mean.
}
\description{
Mean of sparse_numeric
}
