% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_assists.R
\name{most_assists}
\alias{most_assists}
\title{Most Assists}
\usage{
most_assists(year)
}
\arguments{
\item{year}{Specific year to find player with most assists}
}
\description{
This function finds the player with the most assists for a given year
}
\examples{
most_assists(1997)
}
