% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_point_player.R
\name{most_point_player}
\alias{most_point_player}
\title{most_point_player function}
\usage{
most_point_player(year)
}
\arguments{
\item{year}{The year of the data}
}
\description{
This function shows the player with the highest points in the year selected
}
\examples{
most_point_player(year)
}
\keyword{most_point_player}
