% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm for sparse_numeric}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric scalar.
}
\description{
Euclidean Norm (L2) of sparse_numeric
}
