% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Oldest Player}
\usage{
oldest(year)
}
\arguments{
\item{year}{Specific year to find the oldest player}
}
\description{
This function finds the oldest player for a given year
}
\examples{
oldest(1997)
}
