% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{oldest_player function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{The year of the data}
}
\description{
This function shows the oldest player in the year selected
}
\examples{
oldest_player(year)
}
\keyword{oldest_player}
