% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Function2.R
\name{oldest_player_year}
\alias{oldest_player_year}
\title{Function 2: Your input for this function should be an individual year.
Your output should be the oldest player in the NBA data for that year}
\usage{
oldest_player_year(dataset, n)
}
\description{
Function 2: Your input for this function should be an individual year.
Your output should be the oldest player in the NBA data for that year
}
\examples{
oldest_player_year()
}
