% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Overlapping Nonzero Elements}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{First sparse_numeric.}

\item{y}{Second sparse_numeric.}

\item{...}{Additional plotting args passed to plot().}
}
\value{
A scatter plot.
}
\description{
Plot Overlapping Nonzero Elements
}
