% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Show Method for sparse_numeric}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric object.}
}
\description{
Show Method for sparse_numeric
}
