% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Add Two sparse_numeric Vectors}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}
}
\value{
A sparse_numeric object.
}
\description{
Performs elementwise addition of two \code{sparse_numeric} objects.
}
