% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Dot Product for sparse_numeric}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Numeric scalar.
}
\description{
Computes the dot product (crossproduct) of two sparse_numeric vectors.
}
