% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Elementwise Multiplication for sparse_numeric}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (ignored).}

\item{e1}{First operand (for arithmetic operator methods).}

\item{e2}{Second operand (for arithmetic operator methods).}
}
\value{
A sparse_numeric object.
}
\description{
Elementwise Multiplication for sparse_numeric
}
