% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric Vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}
}
\value{
A standardized sparse_numeric object.
}
\description{
Standardize a sparse_numeric Vector
}
