% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Function1.R
\name{top_10_player_pts}
\alias{top_10_player_pts}
\title{Function 1: Your input for this function should be an individual year.
Your output would be the top 10 scored players in that year.}
\usage{
top_10_player_pts(dataset, n)
}
\description{
Function 1: Your input for this function should be an individual year.
Your output would be the top 10 scored players in that year.
}
\examples{
top_10_player_pts()
}
