% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Function3.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Function 3: Your input for this function should be an individual year.
Your output should be the correlation matrix for all numeric variables in the data for that year.}
\usage{
cor_matrix(dataset, n)
}
\description{
Function 3: Your input for this function should be an individual year.
Your output should be the correlation matrix for all numeric variables in the data for that year.
}
\examples{
cor_matrix()
}
