\name{hw5-package}
\alias{hw5-package}
\alias{hw5}
\docType{package}
\title{
\packageTitle{hw5}
}
\description{
\packageDescription{hw5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hw5}
\packageIndices{hw5}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hw5}

Maintainer: \packageMaintainer{hw5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
