% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A Age Function}
\usage{
oldest_player(year)
}
\arguments{
\item{Year}{Who is one of the oldest players in selected year? Defaults to Null}
}
\description{
This function allows you to check one of the oldest players in one specific year
}
